refactor: introduce new account types
