+++ src/routes/LoginWithPersonalAccessToken.tsx
@@ -14,8 +14,12 @@ import { AppContext } from '../context/App';
 import { Constants } from '../utils/constants';
 
 import { Button } from '../components/fields/Button';
-import { getNewTokenURL, isValidHostname, isValidToken } from '../utils/auth';
 import type { AuthOptionsPersonalAccessToken } from '../utils/auth/types';
+import {
+  getNewTokenURL,
+  isValidHostname,
+  isValidToken,
+} from '../utils/auth/utils';
 
 interface IValues {
   token?: string;
